requirements.txt
Pillow==7.1.2
XlsxWriter==1.2.8
